# Conflicts:
#	go.mod
#	go.sum

+++ server/controller/account.go
@@ -1,33 +1,20 @@
 package controller
 
 import (
-	dto "mo2/dto"
-	"mo2/mo2utils"
-
-	"go.mongodb.org/mongo-driver/bson/primitive"
+	"github.com/Monkey-Mouse/mo2/dto"
+	"github.com/Monkey-Mouse/mo2/mo2utils"
 
 	"github.com/gin-gonic/gin"
+	"go.mongodb.org/mongo-driver/bson/primitive"
 
-	//"github.com/swaggo/swag/example/celler/model"
-	"log"
-	"mo2/database"
-	"mo2/server/controller/badresponse"
-	"mo2/server/model"
 	"net/http"
+
+	"github.com/Monkey-Mouse/mo2/database"
+	"github.com/Monkey-Mouse/mo2/server/controller/badresponse"
+	"github.com/Monkey-Mouse/mo2/server/model"
 )
 
 const cookieExpiredTime int = 300000
-
-// @Summary simple test
-// @Description say something
-// @Produce  json
-// @Success 200 {string} json
-// @Router /sayHello [get]
-func SayHello(c *gin.Context) {
-	c.JSON(http.StatusOK, gin.H{
-		"message": "Hello! Welcome to Mo2!",
-	})
-}
 
 // Log godoc
 // @Summary get user info
@@ -35,25 +22,38 @@ func SayHello(c *gin.Context) {
 // @Tags logs
 // @Accept  json
 // @Produce  json
-// @Success 200 {object} dto.LoginUserInfo
+// @Success 200 {object} dto.UserInfo
 // @Router /api/logs [get]
 func (c *Controller) Log(ctx *gin.Context) {
 
 	jwtToken, err := ctx.Cookie("jwtToken")
+	needReset := false
 	var s dto.LoginUserInfo
 	if err != nil {
+		needReset = true
+	} else {
+		//parse jwtToken and get user info
+		s, err = mo2utils.ParseJwt(jwtToken)
+		if err != nil {
+			needReset = true
+		}
+	}
+	if needReset {
 		//allocate an anonymous account
 		account := database.CreateAnonymousAccount()
 		s = dto.Account2SuccessLogin(account)
 		jwtToken = mo2utils.GenerateJwtCode(s)
 		//login success: to record the state
 		ctx.SetCookie("jwtToken", jwtToken, cookieExpiredTime, "/", ctx.Request.Host, false, true)
-	} else {
-		//parse jwtToken and get user info
-		s, err = mo2utils.ParseJwt(jwtToken)
-		if err != nil {
-			log.Println(err)
+	}
+	if dto.Contains(s.Roles, model.OrdinaryUser) {
+		u, ext := database.FindAccountInfo(s.ID)
+		if !ext {
+			ctx.AbortWithStatusJSON(http.StatusUnprocessableEntity, badresponse.SetResponseReason("用户不存在"))
+			return
 		}
+		ctx.JSON(http.StatusOK, u)
+		return
 	}
 	ctx.JSON(http.StatusOK, s)
 }
@@ -66,14 +66,14 @@ func (c *Controller) Log(ctx *gin.Context) {
 // @Produce  json
 // @Param account body model.AddAccountRole true "add new account info"
 // @Success 200 {object} dto.UserInfo
-// @Failure 400 {object} ResponseError
-// @Failure 401 {object} ResponseError
-// @Failure 404 {object} ResponseError
+// @Failure 400 {object} badresponse.ResponseError
+// @Failure 401 {object} badresponse.ResponseError
+// @Failure 404 {object} badresponse.ResponseError
 // @Router /api/accounts/role [post]
 func (c *Controller) AddAccountRole(ctx *gin.Context) {
 	var addAccount model.AddAccountRole
 	if err := ctx.ShouldBindJSON(&addAccount); err != nil {
-		ctx.AbortWithStatusJSON(http.StatusUnauthorized, badresponse.SetResponseError(err))
+		ctx.AbortWithStatusJSON(http.StatusBadRequest, badresponse.SetResponseError(err))
 		return
 	}
 	if err := addAccount.Validation(); err != nil {
@@ -98,14 +98,19 @@ func (c *Controller) AddAccountRole(ctx *gin.Context) {
 // @Produce  json
 // @Param account body dto.UserInfoBrief true "id必须，可修改name/settings"
 // @Success 200 {object} dto.UserInfo
-// @Failure 400 {object} ResponseError
-// @Failure 401 {object} ResponseError
-// @Failure 404 {object} ResponseError
+// @Failure 400 {object} badresponse.ResponseError
+// @Failure 401 {object} badresponse.ResponseError
+// @Failure 404 {object} badresponse.ResponseError
 // @Router /api/accounts [put]
 func (c *Controller) UpdateAccount(ctx *gin.Context) {
 	var accountInfo dto.UserInfoBrief
 	if err := ctx.ShouldBindJSON(&accountInfo); err != nil {
-		ctx.AbortWithStatusJSON(http.StatusUnauthorized, badresponse.SetResponseError(err))
+		ctx.AbortWithStatusJSON(http.StatusBadRequest, badresponse.SetResponseError(err))
+		return
+	}
+	uinfo, _ := mo2utils.GetUserInfo(ctx)
+	if uinfo.ID != accountInfo.ID {
+		ctx.AbortWithStatusJSON(http.StatusForbidden, badresponse.SetResponseReason("非法操作！"))
 		return
 	}
 	account, exist := database.FindAccount(accountInfo.ID)
@@ -118,9 +123,8 @@ func (c *Controller) UpdateAccount(ctx *gin.Context) {
 	if merr := database.UpsertAccount(&account); merr.IsError() {
 		ctx.AbortWithStatusJSON(http.StatusBadRequest, badresponse.SetResponseError(merr))
 		return
-	} else {
-		ctx.JSON(http.StatusOK, dto.Account2UserPublicInfo(account))
 	}
+	ctx.JSON(http.StatusOK, dto.Account2UserPublicInfo(account))
 }
 
 // AddAccount godoc
@@ -131,8 +135,8 @@ func (c *Controller) UpdateAccount(ctx *gin.Context) {
 // @Produce  json
 // @Param account body model.AddAccount true "add new account info"
 // @Success 200 {object} dto.UserInfo
-// @Failure 400 {object} ResponseError
-// @Failure 401 {object} ResponseError
+// @Failure 400 {object} badresponse.ResponseError
+// @Failure 401 {object} badresponse.ResponseError
 // @Router /api/accounts [post]
 func (c *Controller) AddAccount(ctx *gin.Context) {
 	var addAccount model.AddAccount
@@ -141,29 +145,28 @@ func (c *Controller) AddAccount(ctx *gin.Context) {
 		return
 	}
 	if err := addAccount.Validation(); err != nil {
-		ctx.AbortWithStatusJSON(http.StatusBadRequest, badresponse.SetResponseError(err))
+		ctx.AbortWithStatusJSON(http.StatusUnprocessableEntity, badresponse.SetResponseError(err))
 		return
 	}
-	addAccount.UserName = primitive.NewObjectID().Hex() + addAccount.UserName
 	unique, merr := database.EnsureEmailUnique(addAccount.Email)
 	if !unique {
-		ctx.AbortWithStatusJSON(http.StatusBadRequest, badresponse.SetResponseReason("Email已经被使用"))
+		ctx.AbortWithStatusJSON(http.StatusUnprocessableEntity, badresponse.SetResponseReason("Email已经被使用"))
 		return
 	}
 	if merr.IsError() {
-		ctx.AbortWithStatusJSON(http.StatusBadRequest, badresponse.SetResponseError(merr))
+		ctx.AbortWithStatusJSON(http.StatusUnprocessableEntity, badresponse.SetResponseError(merr))
 		return
 	}
-	baseUrl := "http://" + ctx.Request.Host + "/api/accounts/verify"
-	token := mo2utils.GenerateJwtToken(addAccount.Email)
-	url := baseUrl + "?email=" + addAccount.Email + "&token=" + token
-	senderr := mo2utils.SendEmail([]string{addAccount.Email}, mo2utils.VerifyEmailMessage(url, addAccount.UserName), ctx.ClientIP())
+	baseURL := "http://" + ctx.Request.Host + "/api/accounts/verify"
+	token := mo2utils.GenerateVerifyJwtToken(addAccount.Email)
+	url := baseURL + "?email=" + addAccount.Email + "&token=" + token
+	senderr := mo2utils.SendEmail(mo2utils.VerifyEmailMessage(url, addAccount.UserName, []string{addAccount.Email}), ctx.ClientIP())
 	if senderr != nil {
 		ctx.AbortWithStatusJSON(senderr.ErrorCode, badresponse.SetResponseError(senderr))
 	}
 	account, err := database.InitAccount(addAccount, token)
 	if err != nil {
-		ctx.AbortWithStatusJSON(http.StatusUnauthorized, badresponse.SetResponseError(err))
+		ctx.AbortWithStatusJSON(http.StatusUnprocessableEntity, badresponse.SetResponseError(err))
 		return
 	}
 	ctx.JSON(http.StatusOK, dto.Account2UserPublicInfo(account))
@@ -178,9 +181,9 @@ func (c *Controller) AddAccount(ctx *gin.Context) {
 // @Param info body model.DeleteAccount true "delete account info"
 // @Success 202
 // @Success 204
-// @Failure 400 {object} ResponseError
-// @Failure 401 {object} ResponseError
-// @Failure 404 {object} ResponseError
+// @Failure 400 {object} badresponse.ResponseError
+// @Failure 401 {object} badresponse.ResponseError
+// @Failure 404 {object} badresponse.ResponseError
 // @Router /api/accounts [delete]
 func (c *Controller) DeleteAccount(ctx *gin.Context) {
 	var info model.DeleteAccount
@@ -188,15 +191,19 @@ func (c *Controller) DeleteAccount(ctx *gin.Context) {
 		ctx.AbortWithStatusJSON(http.StatusBadRequest, badresponse.SetResponseReason("非法输入"))
 		return
 	}
+	uinfo, _ := mo2utils.GetUserInfo(ctx)
+	if uinfo.Email != info.Email {
+		ctx.AbortWithStatusJSON(http.StatusUnprocessableEntity, badresponse.SetResponseReason("非法输入"))
+		return
+	}
 	if _, err := database.VerifyAccount(model.LoginAccount{Password: info.Password, UserNameOrEmail: info.Email}); err != nil {
-		ctx.AbortWithStatusJSON(http.StatusUnauthorized, badresponse.SetResponseError(err))
+		ctx.AbortWithStatusJSON(http.StatusForbidden, badresponse.SetResponseError(err))
 		return
 	}
 	if _, merr := database.DeleteAccountByEmail(info.Email); merr.IsError() {
-		ctx.Status(http.StatusNoContent)
+		ctx.Status(http.StatusInternalServerError)
 	} else {
-		ctx.Status(http.StatusAccepted)
-
+		ctx.Status(http.StatusNoContent)
 	}
 }
 
@@ -209,7 +216,7 @@ func (c *Controller) DeleteAccount(ctx *gin.Context) {
 // @Param email query string true "email@mo2.com"
 // @Param token query string true "xxxx==sf"
 // @Success 308
-// @Failure 401 {object} ResponseError
+// @Failure 401 {object} badresponse.ResponseError
 // @Router /api/accounts/verify [get]
 func (c *Controller) VerifyEmail(ctx *gin.Context) {
 	var verifyInfo model.VerifyEmail
@@ -235,9 +242,9 @@ func (c *Controller) VerifyEmail(ctx *gin.Context) {
 // @Accept  json
 // @Produce  json
 // @Param account body model.LoginAccount true "login account"
-// @Success 200 {object} dto.LoginUserInfo
-// @Failure 400 {object} ResponseError
-// @Failure 404 {object} ResponseError
+// @Success 200 {object} dto.UserInfo
+// @Failure 400 {object} badresponse.ResponseError
+// @Failure 404 {object} badresponse.ResponseError
 // @Router /api/accounts/login [post]
 func (c *Controller) LoginAccount(ctx *gin.Context) {
 	var loginAccount model.LoginAccount
@@ -246,7 +253,7 @@ func (c *Controller) LoginAccount(ctx *gin.Context) {
 		return
 	}
 	if err := loginAccount.Validation(); err != nil {
-		ctx.AbortWithStatusJSON(http.StatusBadRequest, badresponse.SetResponseError(err))
+		ctx.AbortWithStatusJSON(http.StatusUnprocessableEntity, badresponse.SetResponseError(err))
 		return
 	}
 	account, err := database.VerifyAccount(loginAccount)
@@ -254,8 +261,8 @@ func (c *Controller) LoginAccount(ctx *gin.Context) {
 		ctx.AbortWithStatusJSON(http.StatusUnauthorized, badresponse.SetResponseError(err))
 		return
 	}
-	var s = dto.Account2SuccessLogin(account)
-	jwtToken := mo2utils.GenerateJwtCode(s)
+	var s = dto.Account2UserPublicInfo(account)
+	jwtToken := mo2utils.GenerateJwtCode(dto.Account2SuccessLogin(account))
 	//login success: to record the state
 	ctx.SetCookie("jwtToken", jwtToken, cookieExpiredTime, "/", ctx.Request.Host, false, true)
 	ctx.JSON(http.StatusOK, s)
@@ -283,13 +290,13 @@ func (c *Controller) LogoutAccount(ctx *gin.Context) {
 // @Produce  json
 // @Param id path string false "Account ID"
 // @Success 200 {object} []dto.UserInfo
-// @Failure 400 {object} ResponseError
-// @Failure 404 {object} ResponseError
+// @Failure 400 {object} badresponse.ResponseError
+// @Failure 404 {object} badresponse.ResponseError
 // @Router /api/accounts/detail/{id} [get]
 func (c *Controller) ShowAccount(ctx *gin.Context) {
 	idStr := ctx.Param("id")
 	var us []dto.UserInfo
-	if idStr == "undefined" {
+	if idStr == "undefined" && !mo2utils.IsEnvRelease() {
 		us = database.FindAllAccountsInfo()
 	} else {
 		id, err := primitive.ObjectIDFromHex(idStr)
@@ -314,13 +321,13 @@ func (c *Controller) ShowAccount(ctx *gin.Context) {
 // @Tags accounts
 // @Accept  json
 // @Produce  json
-// @Param userIDs path array true "user IDs list"
+// @Param id query array true "user IDs list"
 // @Success 200 {object} []dto.UserInfoBrief
 // @Router /api/accounts/listBrief [get]
 func (c *Controller) ListAccountsInfo(ctx *gin.Context) {
 	userIDstrs, exist := ctx.GetQueryArray("id")
 	var bs []dto.UserInfoBrief
-	if !exist {
+	if !exist && !mo2utils.IsEnvRelease() {
 		bs = database.ListAllAccountsBrief()
 	} else {
 		bs = database.ListAccountsBrief(userIDstrs)
